fix individual song queuing
