+++ lib/main.dart
@@ -232,14 +232,6 @@ Future<void> initialisation() async {
       await Hive.openBox(boxName);
     }
 
-    currentLikedPlaylistsLength = ValueNotifier<int>(userLikedPlaylists.length);
-    currentLikedSongsLength = ValueNotifier<int>(userLikedSongsList.length);
-    currentOfflineSongsLength = ValueNotifier<int>(userOfflineSongs.length);
-    currentRecentlyPlayedLength = ValueNotifier<int>(userRecentlyPlayed.length);
-    currentLikedAlbumsLength = ValueNotifier<int>(userLikedAlbumsList.length);
-    currentLikedArtistsLength = ValueNotifier<int>(userLikedArtistsList.length);
-    activeQueueLength = ValueNotifier<int>(activeQueue['list'].length);
-
     audioHandler = await AudioService.init(
       builder: ReverbioAudioHandler.new,
       config: const AudioServiceConfig(
@@ -264,6 +256,14 @@ Future<void> initialisation() async {
       }
       userChosenClients = chosenClients;
     }
+
+    currentLikedPlaylistsLength = ValueNotifier<int>(userLikedPlaylists.length);
+    currentLikedSongsLength = ValueNotifier<int>(userLikedSongsList.length);
+    currentOfflineSongsLength = ValueNotifier<int>(userOfflineSongs.length);
+    currentRecentlyPlayedLength = ValueNotifier<int>(userRecentlyPlayed.length);
+    currentLikedAlbumsLength = ValueNotifier<int>(userLikedAlbumsList.length);
+    currentLikedArtistsLength = ValueNotifier<int>(userLikedArtistsList.length);
+    activeQueueLength = ValueNotifier<int>(audioHandler.queueSongBars.length);
 
     try {
       // Listen to incoming links while app is running
